Handle hints files named "0"
If a hints file named "0" was given on the command-line, it wasn’t read.
The code simply checked whether $self->hints is a truthy value; however,
in Perl, the string "0" is falsey.

Checking for definedness instead of truthiness is more robust. It also
helps guard against the case when a user gives --hints "", which is
obviously wrong and should cause zonemaster-cli to error out.