Moves zonemaster-info utility to another repository
This commit removes the zonemaster-info utility from the
Zonemaster-CLI repository. It has, however, already been added
to the Utilities-Using-Zonemaster reposity. It is also under the
Zonemaster Github organization.

This action signals that the utility is not considered to be part
of the core Zonemaster product and that it will not be part of the
Zonemaster release process.