Provide more useful link in man page
The “SEE ALSO” section in the man page had a POD link to a Perl package
called Zonemaster. But no such package actually exists; also, it looks
very awkward when reading the man page version of that documentation.

This commit removes the link that had no chance to work, and replaces it
with something more useful.

As a drive-by change, remove some contractions, as they are usually
frowned upon in written text, unless it is informal.