Slow down the spinner a bit
The console output of zonemaster-cli includes an optional spinner,
looping over four frames, with each frame being advanced whenever
Zonemaster::Engine outputs a message, even if it isn’t shown.

However, over the course of a single domain test, Zonemaster::Engine
outputs thousands of messages of level DEBUG1 or lower. As a result, the
spinner’s animation is far too fast to be enjoyable and a lot of strain
is put on the user’s terminal.

This change simply ensures that no frame is shown for less than 0.1
seconds. The animation looks smoother and doesn’t run the risk
overloading the terminal.